--- src/app/core/components/delete-account-toaster/delete-account-toaster-content/delete-account-toaster-content.component.html
@@ -1,9 +1,0 @@
-<span class="toaster-content-container">
-    <span class="toaster-content-icons">
-        <span class="toaster-content-close-btn" (click)="closeHandler()">
-            <mat-icon aria-label="circle-check">highlight_off_outlined</mat-icon>
-        </span>
-        <mat-icon aria-label="circle-check" class="toaster-content-check-circle"> check_circle_outlined </mat-icon>
-    </span>
-    Your account successfully deleted
-</span>
