+++ src/app/core/components/delete-account-toaster/delete-account-toaster.component.html
@@ -1,7 +1,9 @@
-<dialog-template
-    [dialogHeading]="dialogContent.dialogHeading"
-    [dialogContent]="dialogContent.dialogContent"
-    [handleClose]="openToaster"
->
-    <ng-content></ng-content>
-</dialog-template>
+<span class="toaster-container">
+    <span class="toaster-icons">
+        <span class="toaster-close-btn" (click)="closeHandler()">
+            <mat-icon aria-label="circle-check">highlight_off_outlined</mat-icon>
+        </span>
+        <mat-icon aria-label="circle-check" class="toaster-check-circle"> check_circle_outlined </mat-icon>
+    </span>
+    Your account successfully deleted
+</span>
