+++ src/app/core/components/index.ts
@@ -3,4 +3,3 @@ export * from './sidenav/sidenav.component';
 export * from './notifications/notifications.component';
 export * from './user-profile/user-profile.component';
 export * from './delete-account-toaster/delete-account-toaster.component';
-export * from './delete-account-toaster/delete-account-toaster-content/delete-account-toaster-content.component';
