+++ src/app/core/services/index.ts
@@ -1,3 +1,4 @@
 export * from './user/user.service';
 export * from './access-token/access-token.service';
 export * from './environment/environment.service';
+export * from './user-delete/user-delete.service';
