+++ src/app/shared/dialog/dialog.component.html
@@ -1,1 +1,18 @@
-<span (click)="openDialog()"><ng-content></ng-content></span>
+<div class="dialog-text">
+    <div mat-dialog-title class="dialog-heading">{{ dialogHeading }}</div>
+    <div class="dialog-content">{{ dialogContent }}</div>
+</div>
+<div mat-dialog-actions class="dialog-buttons">
+    <button mat-button mat-dialog-close class="dialog-confirm-btn" (click)="handleClose(dialogEnum.confirm)">
+        Yes
+    </button>
+    <button
+        mat-button
+        mat-dialog-close
+        cdkFocusInitial
+        class="dialog-cancel-btn"
+        (click)="handleClose(dialogEnum.refuse)"
+    >
+        No
+    </button>
+</div>
