+++ src/app/shared/dialog/dialog.component.scss
