+++ src/app/shared/dialog/index.ts
@@ -1,2 +1,1 @@
 export * from './dialog.component';
-export * from './dialog-content/dialog-content.component';
