+++ src/app/shared/index.ts
@@ -2,4 +2,3 @@ export * from './interfaces/environment.interface';
 export * from './material.module';
 export * from './shared.module';
 export * from './dialog/dialog.component';
-export * from './dialog/dialog-content/dialog-content.component';
